# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ExamenTAP API",
        "description": "Documentación de la API REST del backend ExamenTAP.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "API server"
        }
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Inicia sesión con correo electrónico y contraseña",
                "description": "Valida que el usuario exista en Firebase Firestore y que la contraseña sea correcta. Si las credenciales son válidas, genera un token de acceso temporal (Bearer) que debe enviarse en el header Authorization de las siguientes peticiones protegidas.",
                "operationId": "authLogin",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Inicio de sesión exitoso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthLoginResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Credenciales incorrectas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación de los datos enviados",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Demasiados intentos de inicio de sesión",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Cierra la sesión del usuario autenticado",
                "description": "Invalida el token de acceso actual, eliminándolo del almacenamiento en Firebase Firestore. El cliente debe descartar el token localmente tras la respuesta.",
                "operationId": "authLogout",
                "responses": {
                    "200": {
                        "description": "Sesión cerrada correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiSuccessResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado o sesión expirada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/auth/me": {
            "get": {
                "tags": [
                    "Auth"
                ],
                "summary": "Obtiene el usuario autenticado a partir del token actual",
                "description": "Permite al frontend validar y mantener la sesión activa tras recargar la aplicación, comprobando que el token Bearer siga siendo válido.",
                "operationId": "authMe",
                "responses": {
                    "200": {
                        "description": "Sesión activa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AuthMeResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado o sesión expirada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "AuthLoginResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Inicio de sesión exitoso."
                    },
                    "data": {
                        "properties": {
                            "user": {
                                "$ref": "#/components/schemas/User"
                            },
                            "token": {
                                "type": "string",
                                "example": "5f3c9b1e8a7d6c2f4b0a9e8d7c6b5a4f3e2d1c0b9a8f7e6d5c4b3a2f1e0d9c8b"
                            },
                            "token_type": {
                                "type": "string",
                                "example": "Bearer"
                            },
                            "expires_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2026-08-05T13:00:00+00:00"
                            },
                            "permisos": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/SeccionPermiso"
                                }
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "AuthMeResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Sesión activa."
                    },
                    "data": {
                        "properties": {
                            "user": {
                                "$ref": "#/components/schemas/User"
                            },
                            "permisos": {
                                "type": "array",
                                "items": {
                                    "$ref": "#/components/schemas/SeccionPermiso"
                                }
                            }
                        },
                        "type": "object"
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "SeccionPermiso": {
                "properties": {
                    "seccion": {
                        "type": "string",
                        "example": "productos"
                    },
                    "crear": {
                        "type": "boolean",
                        "example": true
                    },
                    "consultar": {
                        "type": "boolean",
                        "example": true
                    },
                    "editar": {
                        "type": "boolean",
                        "example": true
                    },
                    "eliminar": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
            "ApiSuccessResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": true
                    },
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Operación exitosa."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ApiErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Las credenciales proporcionadas son incorrectas."
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ApiValidationErrorResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "status": {
                        "type": "string",
                        "example": "error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation error"
                    },
                    "data": {
                        "type": "object",
                        "example": null,
                        "nullable": true
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "El correo electrónico es obligatorio."
                            ]
                        }
                    }
                },
                "type": "object"
            },
...
            "LoginRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@example.com"
                    },
                    "password": {
                        "type": "string",
                        "format": "password",
                        "example": "Secret123",
                        "minLength": 6
                    }
                },
                "type": "object"
            },
...
            "User": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "f3b1c2d4e5f6"
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan.perez@example.com"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}